requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
Django==3.1.5
django-bootstrap4==2.3.1
django-environ==0.4.5
django-filter~=2.4.0
django-debug-toolbar==3.2
Pillow==8.1.0
pytz==2020.5
soupsieve==2.1
sqlparse==0.4.1